requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Deep Learning
torch==2.2.0
torchvision==0.17.0
timm==0.9.12

# Vector search
faiss-cpu==1.8.0
numpy>=1.26.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pandas==2.1.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
